Miscellaneous Lint-recommended fixes and cleanups

+++ androidapp/app/src/main/java/uk/ac/ncl/team19/lloydsapp/utils/general/CurrencyMangler.java
@@ -9,6 +9,6 @@ public class CurrencyMangler {
     }
 
     public static Long sterlingStringToInteger(String s) {
-        return Long.valueOf((long) (Double.parseDouble(s) * 100));
+        return (long) (Double.parseDouble(s) * 100);
     }
 }
